Remove legacy files, move to lib
