+++ lib/morph.js
