+++ package.json
@@ -33,7 +33,6 @@
     "is-boolean-attribute": "0.0.1",
     "nanoassert": "^1.1.0",
     "nanobench": "^2.1.0",
-    "nanomorph": "git://github.com/tornqvist/nanomorph.git#cache",
     "normalize-html-whitespace": "^0.2.0",
     "through2": "^2.0.3",
     "transform-ast": "^2.4.0"
