+++ tests/browser/raw.js
@@ -1,12 +1,12 @@
 var test = require('tape')
-var html = require('../../')
 var raw = require('../../raw')
+var { html, render } = require('../../')
 
 if (typeof window !== 'undefined') {
   test('unescape html', function (t) {
     t.plan(1)
 
-    var expected = html`<span>Hello there</span>`.toString()
+    var expected = render(html`<span>Hello there</span>`).toString()
     var result = raw('<span>Hello&nbsp;there</span>').toString()
 
     t.equal(expected, result)
